create seperate cgroups when v2 detected
This creates two cgroups: earthly, and buildkit.
The entrypoint.sh, along with all children (e.g. shellrepeater, and
including buildkitd) are moved into the `earthly` cgroup.

buildkitd is configured to spawn new containers inside the `buildkit`
cgroup via a call to runc. When runc starts, it expects the `buildkit`
cgroup container type to be a `domain`; otherwise it hits
https://github.com/opencontainers/runc/blob/d7fc302860dfcd2ae2d266d7cc1ff003d090386f/libcontainer/cgroups/fs2/create.go#L121-L125
which prevents systemd from running in a docker-in-docker setup (as
systemd is expecting a domain type).

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>